requirements.txt
pygetwindow==0.0.9
opencv-python==4.8.1.78
pytesseract==0.3.10
deepl==1.15.0
Pillow==10.0.1
PyQt5==5.15.10
mss==9.0.1
pywin32==306
numpy==1.25.2
